修改id類型

+++ database/migrations/2024_08_23_141125_create_role_has_permissions_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('role_has_permissions', function (Blueprint $table) {
-            $table->Integer('role_id');
-            $table->Integer('permission_id');
+            $table->unsignedBigInteger('role_id');
+            $table->unsignedBigInteger('permission_id');
             
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');#設定外鍵
             $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
